requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appnope==0.1.0            # via ipython
attrs==19.2.0             # via jsonschema
backcall==0.1.0           # via ipython
bleach==3.1.0             # via nbconvert
click==7.0                # via pip-tools
cycler==0.10.0            # via matplotlib
decorator==4.4.0          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via nbconvert
ipykernel==5.1.2          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via jupyter-contrib-nbextensions, nbformat, notebook, qtconsole, traitlets
ipython==7.8.0            # via ipykernel, ipywidgets, jupyter-console, jupyter-latex-envs
ipywidgets==7.5.1         # via jupyter
jedi==0.15.1              # via ipython
jinja2==2.10.3            # via jupyterlab, jupyterlab-server, nbconvert, notebook
json5==0.8.5              # via jupyterlab-server
jsonschema==3.0.2         # via jupyterlab-server, nbformat
jupyter-client==5.3.4     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==6.0.0    # via jupyter
jupyter-contrib-core==0.3.3  # via jupyter-contrib-nbextensions, jupyter-nbextensions-configurator
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.0       # via jupyter-client, jupyter-contrib-core, jupyter-contrib-nbextensions, jupyter-latex-envs, jupyter-nbextensions-configurator, nbconvert, nbformat, notebook, qtconsole
jupyter-highlight-selected-word==0.2.0  # via jupyter-contrib-nbextensions
jupyter-latex-envs==1.4.6  # via jupyter-contrib-nbextensions
jupyter-nbextensions-configurator==0.4.1  # via jupyter-contrib-nbextensions
jupyter==1.0.0
jupyterlab-server==1.0.6  # via jupyterlab
jupyterlab==1.1.4
kiwisolver==1.1.0         # via matplotlib
lxml==4.4.1               # via jupyter-contrib-nbextensions
markupsafe==1.1.1         # via jinja2
matplotlib==3.1.1
mistune==0.8.4            # via nbconvert
nbconvert==5.6.0          # via jupyter, jupyter-contrib-nbextensions, jupyter-latex-envs, notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==6.0.1           # via jupyter, jupyter-contrib-core, jupyter-contrib-nbextensions, jupyter-latex-envs, jupyter-nbextensions-configurator, jupyterlab, jupyterlab-server, widgetsnbextension
numpy==1.17.2
pandas==0.25.1
pandocfilters==1.4.2      # via nbconvert
parso==0.5.1              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==6.2.0
pip-tools==4.1.0
prometheus-client==0.7.1  # via notebook
prompt-toolkit==2.0.10    # via ipython, jupyter-console
psycopg2-binary==2.8.3
ptyprocess==0.6.0         # via pexpect, terminado
pygments==2.4.2           # via ipython, jupyter-console, nbconvert, qtconsole
pyparsing==2.4.2          # via matplotlib
pyrsistent==0.15.4        # via jsonschema
python-dateutil==2.8.0    # via jupyter-client, matplotlib, pandas
pytz==2019.3              # via pandas
pyyaml==5.1.2             # via jupyter-contrib-nbextensions, jupyter-nbextensions-configurator
pyzmq==18.1.0             # via jupyter-client, notebook
qtconsole==4.5.5          # via jupyter
send2trash==1.5.0         # via notebook
six==1.12.0               # via bleach, cycler, jsonschema, pip-tools, prompt-toolkit, pyrsistent, python-dateutil, torchvision, traitlets
terminado==0.8.2          # via notebook
testpath==0.4.2           # via nbconvert
torch==1.2.0
torchvision==0.4.0
tornado==6.0.3            # via ipykernel, jupyter-client, jupyter-contrib-core, jupyter-contrib-nbextensions, jupyter-nbextensions-configurator, jupyterlab, notebook, terminado
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-contrib-core, jupyter-contrib-nbextensions, jupyter-core, jupyter-latex-envs, jupyter-nbextensions-configurator, nbconvert, nbformat, notebook, qtconsole
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.4.0        # via ipython, jsonschema, jupyter-contrib-core, kiwisolver
